Finalize all search algorithms

+++ src/code/search_queues/AStarQueue.java
@@ -3,16 +3,14 @@ package code.search_queues;
 import code.Node;
 import code.heuristics.Heuristic;
 
-import java.util.Comparator;
-import java.util.PriorityQueue;
-import java.util.Queue;
+import java.util.*;
 
 public class AStarQueue implements SearchQueue {
 
     private final Queue<Node> queue;
 
     public AStarQueue(Heuristic heuristic) {
-        this.queue = new PriorityQueue<>(Comparator.comparingInt(a -> a.getState().getMoneySpent() + heuristic.apply(a)));
+        this.queue = new PriorityQueue<>(Comparator.comparingDouble(a -> a.getState().getMoneySpent() + heuristic.apply(a)));
     }
 
     @Override
